+++ modules/gateway-task/main.tf
@@ -35,6 +35,7 @@ locals {
       partition                 = var.consul_partition
       primary_datacenter        = var.consul_primary_datacenter
       enable_token_replication  = var.enable_acl_token_replication
+      audit_logging             = var.audit_logging
     }
   )