+++ modules/mesh-task/variables.tf
@@ -429,3 +429,9 @@ variable "consul_ecs_config" {
     ]))
   }
 }
+
+variable "audit_logging" {
+  description = "Whether to enable audit logging for the Consul agent [Consul Enterprise]. ACLs must be enabled to enable audit logging."
+  type        = bool
+  default     = false
+}