+++ test/acceptance/tests/hcp/terraform/hcp-install/variables.tf
@@ -91,3 +91,9 @@ variable "consul_ca_cert_secret_arn" {
   description = "ARN of the secret holding the Consul CA certificate."
   type        = string
 }
+
+variable "audit_logging" {
+  description = "Whether audit logging is enabled."
+  type        = bool
+  default     = false
+}